requirements.txt
streamlit==1.42.2
owlready2==0.47
mirascope==1.19.0
mirascope[openai]==1.19.0
python-dotenv==1.0.1
pydantic==2.10.6
python-i18n[YAML]==0.3.9